Refactor tests
Remove all hard-coded sleeps **except** the last one, when we delete the
csi-lvm-controller, otherwise PVCs may not get deleted before the
controller is deleted. When this happens, the loop devices will not be
cleared correctly when running `make test-cleanup`.

We also want to test one thing per test, otherwise we may not know why a
test failed. We leverage `kubectl wait --for=jsonpath=` as much as
possible. This way the tests do not need to check for specific strings,
we let `--for=jsonpath=` do that. The best part with this approach is
that we can use the `--timeout` flag. This brings the **entire**
integration test suite duration to 70 seconds. Before this change, the
sleeps alone (170s) would take longer than that.

To double-check for race conditions or flaky tests, I ran all tests
locally 100 times with `RERUN=100 make test`. All 100 runs passed. This
looks good to me!

Separately, I have also tested this in Talos v1.4.0 running K8s 1.26.4.
Everything works as expected now. See this PR comment for more details:
https://github.com/metal-stack/csi-driver-lvm/pull/87#issuecomment-1519667687

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>