Add support for Talos OS (#87)
* Add Talos OS support

`/etc` on Talos OS is read-only, which means that new PVs will fail to
create. This change makes the hard-coded `/etc/lvm` configurable via the
`--hostwritepath` flag.

NOTE that this also changes the current `/run/lock/lvm` to
`/etc/lvm/lock`.

This is a requirement for
https://github.com/metal-stack/helm-charts/pull/64

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>

* Create loop devices part of `make test`

After this change, people which expect integration tests to be
self-contained, will not be disappointed. It took me a while to figure
out why **some** integration tests were failing locally. I eventually
found out about this requirement in this doc page:
https://docs.metal-stack.io/stable/external/csi-driver-lvm/README/. The
GitHub Actions workflow also helped. Even then, the mknod command was
not mentioned anywhere. My NixOS host did not have these special files
/dev/loop100 & /dev/loop101 created. With this change, `make test` is
self-contained & it should work the same on all Linux hosts, whether
it's a local development workstation or running in GitHub Actions.

Speaking of GitHub Actions, we do not want to run the build-platforms
job if the DOCKER_REGISTRY_TOKEN secret is not set. If we don't check
for this, the job will fail in repo forks, where these secrets will not
be available by default. FWIW, `${{ secrets. }}` is not available in
`if` conditions. The secret value needs to be exposed as an env for the
`if` condition to work correctly. FTR:
https://github.com/orgs/community/discussions/26726

I also remembered to remove the loop devices part of `make
test-cleanup` & double-check that the loop device has been actually
removed. I have hit a situation where the file was deleted, but
/dev/loop100 was still left dangling. Had to `sudo dmsetup remove` it.

Lastly, Docker CLI is configured to ignore the *.img files. These are
created in the same directory and should not be sent to Docker when
running `docker build`.

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>

* Refactor tests

Remove all hard-coded sleeps **except** the last one, when we delete the
csi-lvm-controller, otherwise PVCs may not get deleted before the
controller is deleted. When this happens, the loop devices will not be
cleared correctly when running `make test-cleanup`.

We also want to test one thing per test, otherwise we may not know why a
test failed. We leverage `kubectl wait --for=jsonpath=` as much as
possible. This way the tests do not need to check for specific strings,
we let `--for=jsonpath=` do that. The best part with this approach is
that we can use the `--timeout` flag. This brings the **entire**
integration test suite duration to 70 seconds. Before this change, the
sleeps alone (170s) would take longer than that.

To double-check for race conditions or flaky tests, I ran all tests
locally 100 times with `RERUN=100 make test`. All 100 runs passed. This
looks good to me!

Separately, I have also tested this in Talos v1.4.0 running K8s 1.26.4.
Everything works as expected now. See this PR comment for more details:
https://github.com/metal-stack/csi-driver-lvm/pull/87#issuecomment-1519667687

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>

---------

Signed-off-by: Gerhard Lazu <gerhard@lazu.ch>